+++ app/src/main/java/ru/anydevprojects/simplepodcastapp/podcastEpisode/domain/PodcastEpisodeRepository.kt
@@ -4,9 +4,9 @@ import ru.anydevprojects.simplepodcastapp.podcastEpisode.domain.models.PodcastEp
 
 interface PodcastEpisodeRepository {
 
-    suspend fun getEpisodesByPodcastId(podcastId: Int): Result<List<PodcastEpisode>>
+    suspend fun getEpisodesByPodcastId(podcastId: Long): Result<List<PodcastEpisode>>
 
     suspend fun getEpisodesByPodcastIds(podcastIds: List<Int>): Result<List<PodcastEpisode>>
 
-    suspend fun getEpisodeById(id: Int): Result<PodcastEpisode>
+    suspend fun getEpisodeById(id: Long): Result<PodcastEpisode>
 }
