+++ app/src/main/java/ru/anydevprojects/simplepodcastapp/podcastEpisode/domain/models/PodcastEpisode.kt
@@ -1,14 +1,14 @@
 package ru.anydevprojects.simplepodcastapp.podcastEpisode.domain.models
 
 data class PodcastEpisode(
-    val id: Int,
+    val id: Long,
     val title: String,
     val description: String,
-    val dateTimestamp: Int,
+    val dateTimestamp: Long,
     val enclosureUrl: String,
     val enclosureType: String,
     val enclosureLength: Int,
     val duration: Int?,
     val feedImage: String,
-    val feedId: Int
+    val feedId: Long
 )
