+++ app/src/main/java/ru/anydevprojects/simplepodcastapp/podcastEpisode/presentation/PodcastEpisodeScreenNavigation.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PodcastEpisodeScreenNavigation(
     val podcastName: String,
-    val episodeId: Int
+    val episodeId: Long
 )
